+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - help for rake:update
 - clean panel with zero scores
 - spin level: B is 0 ??
+- accept_categories: parser or updater?
+- json:: depends to serverside ??
+- limit: spec
  
 ### new feature